requirements.txt
fastapi
uvicorn
onnxruntime
pillow
numpy
python-multipart  
torch
torchvision
transformers
requests
